+++ mstrio/__init__.py
@@ -23,8 +23,10 @@ administrative workflows for maintaining a Strategy One environment.
 """
 
 __title__ = "mstrio-py"
-__version__ = "11.5.3.101"  # NOSONAR
+__version__ = "11.5.4.101"  # NOSONAR
 __license__ = "Apache License 2.0"
 __description__ = "Python interface for the Strategy One REST API"
 __author__ = "Strategy One"
 __author_email__ = "bkaczynski@strategy.com"
+
+from .utils.dev_helpers import what_can_i_do_with  # noqa: F401